update: refactored the code based on the PR review
