+++ apps/web/package.json
@@ -24,8 +24,10 @@
         "next-themes": "^0.3.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
+        "react-hook-form": "^7.51.3",
         "sonner": "^1.4.41",
-        "ui": "workspace:*"
+        "ui": "workspace:*",
+        "zod": "link:@hookform/resolvers/zod"
     },
     "devDependencies": {
         "@100mslive/react-sdk": "^0.10.10",
